Move EnvironmentExtensions.cs into the Extensions folder
